GenQuestions Working

+++ QB/mark.c
@@ -16,7 +16,7 @@ char *C_A = "c/answerset_c.csv"; //qb ans for c
 * modifies sending string to the serialised string to be sent to TM
 * sending string: length//;<t/f>//;(if is last attempt)expected_answer//;student_answer
 */
-struct FileData question_correct(uint32_t seed, char index,char lastAttempt,char*answer){
+struct FileData question_correct(uint64_t seed, char index,char lastAttempt,char*answer){
     int*ids = malloc((index+1)*sizeof(int));
     question_ids(ids,index+1,seed);
     char *sep = ","; //general and ans seperator
